Update to swift 5.0. Tested on Ubuntu and macOS
