--- Sources/CGtk/header.h
@@ -1,5 +1,0 @@
-//
-// Created by GothStar on 23.03.19.
-//
-
-#include <termios.h>
