--- Sources/CGtk/module.modulemap
@@ -1,20 +1,0 @@
-
-module CGtk [system] {
-
- header "header.h"
-  header "/usr/include/gtk-3.0/gtk/gtk.h"
-
-  link "gtk-3"
-  link "gdk-3"
-  link "pangocairo-1.0"
-  link "pango-1.0"
-  link "atk-1.0"
-  link "cairo-gobject"
-  link "cairo"
-  link "gdk_pixbuf-2.0"
-  link "gio-2.0"
-  link "gobject-2.0"
-  link "glib-2.0"
-
-  export *
-  }
