+++ Sources/sGtk/ApplicationWindow.swift
