+++ Sources/sGtk/Common.swift
