+++ Sources/sGtk/Image.swift
