+++ Sources/sGtk/PositionType.swift
