+++ Sources/sGtk/SignalBox.swift
