+++ Sources/sGtk/Signals.swift
