+++ Sources/sGtk/UnsafePointerGChar+ToString.swift
