+++ Sources/sGtk/Widget+CastedPointer.swift
