--- SwiftGtkDemo/Package@swift-4.2.swift
@@ -1,16 +1,0 @@
-// swift-tools-version:4.2
-
-import PackageDescription
-
-let package = Package(
-        name: "SwiftGtkDemo",
-        dependencies: [
-            .package(url: "https://github.com/GothStar/SwiftGtk.git", .branch("master"))
-        ],
-        targets: [
-
-            .target(name: "SwiftGtkDemo", dependencies: ["sGtk"], path: "Sources/SwiftGtkDemo"),
-            .testTarget(name: "SwiftGtkDemoTest", dependencies: ["SwiftGtkDemo"])
-
-        ]
-)
