+++ SwiftGtkDemo/Resources/GTK.png
